// src/effects/combined/CombinedOut.d.ts
import ExitAnimationEffect from "../../core/effects/ExitAnimationEffect";
import React from "react";
export declare class CombinedOut implements ExitAnimationEffect {
    style: React.CSSProperties;
    duration: number;
    constructor(animations: ExitAnimationEffect[]);
}
//# sourceMappingURL=CombinedOut.d.ts.map