// src/effects/expand/Shrink.d.ts
import Easing from "../../core/easing/Easing";
import ExitAnimationEffect from "../../core/effects/ExitAnimationEffect";
import React from "react";
export declare class Shrink implements ExitAnimationEffect {
    style: React.CSSProperties;
    duration: number;
    constructor(targetWidth?: number, targetHeight?: number, duration?: number, easing?: Easing);
}
//# sourceMappingURL=Shrink.d.ts.map