// src/effects/fade/FadeOut.d.ts
import Easing from "../../core/easing/Easing";
import React from "react";
import ExitAnimationEffect from "../../core/effects/ExitAnimationEffect";
export declare class FadeOut implements ExitAnimationEffect {
    style: React.CSSProperties;
    duration: number;
    constructor(targetAlpha?: number, duration?: number, easing?: Easing);
}
//# sourceMappingURL=FadeOut.d.ts.map