// src/effects/slide/SlideOut.d.ts
import Easing from "../../core/easing/Easing";
import ExitAnimationEffect from "../../core/effects/ExitAnimationEffect";
import React from "react";
export declare class SlideOut implements ExitAnimationEffect {
    style: React.CSSProperties;
    duration: number;
    constructor(xPercentage?: number, yPercentage?: number, duration?: number, easing?: Easing);
}
//# sourceMappingURL=SlideOut.d.ts.map